+++ dzVents/runtime/Variable.lua
@@ -13,10 +13,10 @@ local function Variable(domoticz, data)
 		['baseType'] = domoticz.BASETYPE_VARIABLE,
 		isVariable = true,
 		isHTTPResponse = false,
-	    isDevice = false,
-	    isScene = false,
-	    isGroup = false,
-	    isTimer = false,
+		isDevice = false,
+		isScene = false,
+		isGroup = false,
+		isTimer = false,
 		isSecurity = false
 	}
 
@@ -61,6 +61,15 @@ local function Variable(domoticz, data)
 			type = 'variable',
 			idx = data.id
 		})
+	end
+
+	function self.rename(newName)
+		local url = domoticz.settings['Domoticz url'] .. '/json.htm?type=command&param=updateuservariable' ..
+					'&idx=' .. data.id ..
+					'&vname=' .. domoticz.utils.urlEncode(newName) ..
+					'&vtype=' ..  self.type ..
+					'&vvalue=' .. domoticz.utils.urlEncode(self.value)
+		domoticz.openURL(url)
 	end
 
 	return self