+++ dzVents/runtime/device-adapters/generic_device.lua
@@ -43,15 +43,16 @@ return {
 	baseType = 'device',
 
 	match = function (device)
-		return true -- generic always matches
+		return true -- generic always matches?
 	end,
 
 	matches = function (device, adapterManager)
 		adapterManager.addDummyMethod(device, 'setDescription')
 		adapterManager.addDummyMethod(device, 'setIcon')
 		adapterManager.addDummyMethod(device, 'setValues')
+		adapterManager.addDummyMethod(device, 'rename')
 	end,
-	
+
 	process = function (device, data, domoticz, utils, adapterManager)
 		local _states = adapterManager.states
 
@@ -133,20 +134,29 @@ return {
 
 		function device.setDescription(description)
 			local url = domoticz.settings['Domoticz url'] ..
-				"/json.htm?description=" .. domoticz.utils.urlEncode(description) ..
+				"/json.htm?description=" .. utils.urlEncode(description) ..
 				"&idx=" .. device.id ..
-				"&name=".. domoticz.utils.urlEncode(device.name) ..
+				"&name=".. utils.urlEncode(device.name) ..
 				"&type=setused&used=true"
 			return domoticz.openURL(url)
 		end
-		
+
 		function device.setIcon(iconNumber)
 			local url = domoticz.settings['Domoticz url'] .. 
-				'/json.htm?type=setused&used=true&name=' .. domoticz.utils.urlEncode(device.name) ..
-				'&description=' .. domoticz.utils.urlEncode(device.description) ..
+				'/json.htm?type=setused&used=true&name=' .. 
+				 utils.urlEncode(device.name) ..
+				'&description=' .. utils.urlEncode(device.description) ..
 				'&idx=' .. device.id .. 
 				'&switchtype=' .. device.switchTypeValue ..
 				'&customimage=' .. iconNumber
+			return domoticz.openURL(url)
+		end
+
+		function device.rename(newName)
+			local url = domoticz.settings['Domoticz url'] ..  
+						"/json.htm?type=command&param=renamedevice" ..
+						"&idx=" .. device.idx ..
+						"&name=" .. utils.urlEncode(newName)
 			return domoticz.openURL(url)
 		end