+++ hardware/OpenZWave.cpp
@@ -1688,28 +1688,36 @@ void COpenZWave::AddValue(const OpenZWave::ValueID& vID, const NodeInfo* pNodeIn
 	}
 	else if ((commandclass == COMMAND_CLASS_ALARM) || (commandclass == COMMAND_CLASS_SENSOR_ALARM))
 	{
-		if (vType != OpenZWave::ValueID::ValueType_List)
+		if (
+			(vType != OpenZWave::ValueID::ValueType_List) &&
+			(vType != OpenZWave::ValueID::ValueType_Byte)
+			)
 		{
 			_log.Log(LOG_ERROR, "OpenZWave: Unhandled value type: %d, %s:%d", vType, std::string(__MYFUNCTION__).substr(std::string(__MYFUNCTION__).find_last_of("/\\") + 1).c_str(), __LINE__);
 			return;
 		}
+		int32 alarm_value = 0;
+		if (vType == OpenZWave::ValueID::ValueType_List)
+		{
+			int32 listValue = 0;
+			if (m_pManager->GetValueListSelection(vID, &listValue) == false)
+				return;
+			alarm_value = listValue;
+		}
+		else if (vType == OpenZWave::ValueID::ValueType_Byte)
+		{
+			if (m_pManager->GetValueAsByte(vID, &byteValue) == false)
+				return;
+			alarm_value = 0; //we assume a previous event cleared message was received
+		}
+
 		int newInstance = GetIndexFromAlarm(vLabel);
 		if (newInstance == 0)
 			newInstance = vOrgIndex;
 		_device.instanceID = newInstance;
-		//std::vector<std::string> tModes;
-		//m_pManager->GetValueListItems(vID, &tModes);
-
-		int32 listValue = 0;
-		if (m_pManager->GetValueListSelection(vID, &listValue))
-		{
-			_device.devType = ZDTYPE_SWITCH_NORMAL;
-			if (listValue == 0) //clear
-				_device.intvalue = 0;
-			else
-				_device.intvalue = 255;
-			InsertDevice(_device);
-		}
+		_device.devType = ZDTYPE_SWITCH_NORMAL;
+		_device.intvalue = (alarm_value == 0) ? 0 : 255;
+		InsertDevice(_device);
 	}
 	else if (commandclass == COMMAND_CLASS_METER)
 	{
@@ -1799,14 +1807,21 @@ void COpenZWave::AddValue(const OpenZWave::ValueID& vID, const NodeInfo* pNodeIn
 			return;
 
 		if (
-			(vOrgIndex == ValueID_Index_SensorMultiLevel::Temperature)
-			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::WaterTemperature)
-			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::SoilTemperature)
-			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::TargetTemperature)
-			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::BoilerWaterTemperature)
-			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::DHWTemperature)
-			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::OutsideTemperature)
-			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::ExhaustTemperature)
+			(vOrgIndex == ValueID_Index_SensorMultiLevel::Air_Temperature)
+			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::Water_Temperature)
+			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::Soil_Temperature)
+			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::Target_Temperature)
+			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::Boiler_Water_Temperature)
+			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::Domestic_Hot_Water_Temperature)
+			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::Outside_Temperature)
+			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::Exhaust_Temperature)
+			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::Return_Air_Temperature)
+			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::Supply_Air_Temperature)
+			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::Condenser_Coil_Temperature)
+			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::Evaporator_Coil_Temperature)
+			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::Liquid_Line_Temperature)
+			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::Discharge_Line_Temperature)
+			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::Defrost_Temperature)
 			) 
 		{
 			if (vUnits == "F")
@@ -1831,8 +1846,8 @@ void COpenZWave::AddValue(const OpenZWave::ValueID& vID, const NodeInfo* pNodeIn
 			_device.devType = ZDTYPE_SENSOR_LIGHT;
 		}
 		else if (
-			(vOrgIndex == ValueID_Index_SensorMultiLevel::RelativeHumidity)
-			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::SoilHumidity)
+			(vOrgIndex == ValueID_Index_SensorMultiLevel::Humidity)
+			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::Soil_Humidity)
 			)
 		{
 			_device.intvalue = round(fValue);
@@ -1847,14 +1862,14 @@ void COpenZWave::AddValue(const OpenZWave::ValueID& vID, const NodeInfo* pNodeIn
 			_device.devType = ZDTYPE_SENSOR_VELOCITY;
 		}
 		else if (
-			(vOrgIndex == ValueID_Index_SensorMultiLevel::BarometricPressure)
-			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::AtmosphericPressure)
+			(vOrgIndex == ValueID_Index_SensorMultiLevel::Barometric_Pressure)
+			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::Atmospheric_Pressure)
 			)
 		{
 			_device.scaleMultiply = 10.0f;
 			_device.devType = ZDTYPE_SENSOR_BAROMETER;
 		}
-		else if (vOrgIndex == ValueID_Index_SensorMultiLevel::DewPoint)
+		else if (vOrgIndex == ValueID_Index_SensorMultiLevel::Dew_Point)
 		{
 			_device.devType = ZDTYPE_SENSOR_DEWPOINT;
 		}
@@ -1882,7 +1897,10 @@ void COpenZWave::AddValue(const OpenZWave::ValueID& vID, const NodeInfo* pNodeIn
 		{
 			_device.devType = ZDTYPE_SENSOR_WATER;
 		}
-		else if (vOrgIndex == ValueID_Index_SensorMultiLevel::CO2)
+		else if (
+			(vOrgIndex == ValueID_Index_SensorMultiLevel::Carbon_Dioxide)
+			|| (vOrgIndex == ValueID_Index_SensorMultiLevel::Carbon_Monoxide)
+			)
 		{
 			_device.devType = ZDTYPE_SENSOR_CO2;
 		}
@@ -1890,20 +1908,20 @@ void COpenZWave::AddValue(const OpenZWave::ValueID& vID, const NodeInfo* pNodeIn
 		{
 			_device.devType = ZDTYPE_SENSOR_MOISTURE;
 		}
-		else if (vOrgIndex == ValueID_Index_SensorMultiLevel::TankCapacity)
+		else if (vOrgIndex == ValueID_Index_SensorMultiLevel::Tank_Capacity)
 		{
 			_device.devType = ZDTYPE_SENSOR_TANK_CAPACITY;
 		}
-		else if (vOrgIndex == ValueID_Index_SensorMultiLevel::General)
+		else if (vOrgIndex == ValueID_Index_SensorMultiLevel::General_Purpose)
 		{
 			_device.devType = ZDTYPE_SWITCH_NORMAL;
 		}
-		else if (vOrgIndex == ValueID_Index_SensorMultiLevel::RainRate)
+		else if (vOrgIndex == ValueID_Index_SensorMultiLevel::Rain_Rate)
 		{
 			_device.custom_label = "mm/h";
 			_device.devType = ZDTYPE_SENSOR_CUSTOM;
 		}
-		else if (vOrgIndex == ValueID_Index_SensorMultiLevel::SeismicIntensity)
+		else if (vOrgIndex == ValueID_Index_SensorMultiLevel::Seismic_Intensity)
 		{
 			_device.devType = ZDTYPE_SENSOR_PERCENTAGE;
 		}
@@ -4322,17 +4340,11 @@ void COpenZWave::GetNodeValuesJson(const unsigned int homeID, const int nodeID, 
 						std::string i_units = m_pManager->GetValueUnits(ittValue);
 						std::string i_help = m_pManager->GetValueHelp(ittValue);
 
-						struct tm timeinfo;
-						localtime_r(&ittCmds.second.m_LastSeen, &timeinfo);
-
-						char szDate[30];
-						strftime(szDate, sizeof(szDate), "%Y-%m-%d %H:%M:%S", &timeinfo);
-
 						root["result"][index]["config"][ivalue]["index"] = i_index;
 						root["result"][index]["config"][ivalue]["label"] = i_label;
 						root["result"][index]["config"][ivalue]["units"] = i_units;
 						root["result"][index]["config"][ivalue]["help"] = i_help;
-						root["result"][index]["config"][ivalue]["LastUpdate"] = szDate;
+						root["result"][index]["config"][ivalue]["LastUpdate"] = TimeToString(&ittCmds.second.m_LastSeen, TF_DateTime);
 						ivalue++;
 					}
 					else if (commandclass == COMMAND_CLASS_WAKE_UP)
@@ -4353,16 +4365,11 @@ void COpenZWave::GetNodeValuesJson(const unsigned int homeID, const int nodeID, 
 								std::string i_units = m_pManager->GetValueUnits(ittValue);
 								std::string i_help = m_pManager->GetValueHelp(ittValue);
 
-								struct tm timeinfo;
-								localtime_r(&ittCmds.second.m_LastSeen, &timeinfo);
-
-								char szDate[30];
-								strftime(szDate, sizeof(szDate), "%Y-%m-%d %H:%M:%S", &timeinfo);
 								root["result"][index]["config"][ivalue]["index"] = i_index;
 								root["result"][index]["config"][ivalue]["label"] = i_label;
 								root["result"][index]["config"][ivalue]["units"] = i_units;
 								root["result"][index]["config"][ivalue]["help"] = i_help;
-								root["result"][index]["config"][ivalue]["LastUpdate"] = szDate;
+								root["result"][index]["config"][ivalue]["LastUpdate"] = TimeToString(&ittCmds.second.m_LastSeen, TF_DateTime);
 								ivalue++;
 							}
 						}
@@ -4810,12 +4817,7 @@ namespace http {
 						root["result"][ii]["IsPlus"] = pNode->IsPlus;
 						root["result"][ii]["Generic_type"] = pOZWHardware->GetNodeGenericType(pNode->IsPlus, homeID, nodeID);
 
-						char szDate[80];
-						struct tm loctime;
-						localtime_r(&pNode->LastSeen, &loctime);
-						strftime(szDate, 80, "%Y-%m-%d %X", &loctime);
-
-						root["result"][ii]["LastUpdate"] = szDate;
+						root["result"][ii]["LastUpdate"] = TimeToString(&pNode->LastSeen, TF_DateTime);
 
 						//Add configuration parameters here
 						pOZWHardware->GetNodeValuesJson(homeID, nodeID, root, ii);