+++ hardware/openzwave/Utils.h
@@ -36,6 +36,14 @@
 #include <algorithm>
 #include <sstream>
 #include <vector>
+#ifndef WIN32
+#ifndef WINRT
+#ifdef DEBUG
+#include <execinfo.h>
+#include <cxxabi.h>
+#endif
+#endif
+#endif
 
 namespace OpenZWave
 {
@@ -147,13 +155,113 @@ namespace OpenZWave
 		string intToString(int x);
 
 		const char* rssi_to_string(uint8 _data);
+
+#ifndef WIN32
+#ifndef WINRT
+#ifdef DEBUG
+		class StackTraceGenerator
+		{
+		private:
+
+		    //  this is a pure utils class
+		    //  cannot be instantiated
+		    //
+		    StackTraceGenerator() = delete;
+		    StackTraceGenerator(const StackTraceGenerator&) = delete;
+		    StackTraceGenerator& operator=(const StackTraceGenerator&) = delete;
+		    ~StackTraceGenerator() = delete;
+
+		public:
+
+		    static std::vector<std::string> GetTrace()
+		    {
+		        //  record stack trace upto 128 frames
+		        int callstack[128] = {};
+
+		        // collect stack frames
+		        int  frames = backtrace((void**) callstack, 5);
+
+		        // get the human-readable symbols (mangled)
+		        char** strs = backtrace_symbols((void**) callstack, frames);
+
+		        std::vector<std::string> stackFrames;
+		        stackFrames.reserve(frames);
+
+		        for (int i = 2; i < frames; ++i)
+		        {
+		            char functionSymbol[1024] = {};
+		            char moduleName[1024] = {};
+		            int  offset = 0;
+		            char addr[48] = {};
+
+		            /*
+
+		             Typically this is how the backtrace looks like:
+
+		             0   <app/lib-name>     0x0000000100000e98 _Z5tracev + 72
+		             1   <app/lib-name>     0x00000001000015c1 _ZNK7functorclEv + 17
+		             2   <app/lib-name>     0x0000000100000f71 _Z3fn0v + 17
+		             3   <app/lib-name>     0x0000000100000f89 _Z3fn1v + 9
+		             4   <app/lib-name>     0x0000000100000f99 _Z3fn2v + 9
+		             5   <app/lib-name>     0x0000000100000fa9 _Z3fn3v + 9
+		             6   <app/lib-name>     0x0000000100000fb9 _Z3fn4v + 9
+		             7   <app/lib-name>     0x0000000100000fc9 _Z3fn5v + 9
+		             8   <app/lib-name>     0x0000000100000fd9 _Z3fn6v + 9
+		             9   <app/lib-name>     0x0000000100001018 main + 56
+		             10  libdyld.dylib      0x00007fff91b647e1 start + 0
+
+		             */
+
+		            // split the string, take out chunks out of stack trace
+		            // we are primarily interested in module, function and address
+		            sscanf(strs[i], "%*s %s %s %s %*s %d",
+		                   moduleName, addr, functionSymbol, &offset);
+
+		            int   validCppName = 0;
+		            //  if this is a C++ library, symbol will be demangled
+		            //  on success function returns 0
+		            //
+		            char* functionName = abi::__cxa_demangle(functionSymbol,
+		                                                     NULL, 0, &validCppName);
+
+		            char stackFrame[4096] = {};
+		            if (validCppName == 0) // success
+		            {
+		                sprintf(stackFrame, "(%s)\t0x%s — %s + %d",
+		                        moduleName, addr, functionName, offset);
+		            }
+		            else
+		            {
+		                //  in the above traceback (in comments) last entry is not
+		                //  from C++ binary, last frame, libdyld.dylib, is printed
+		                //  from here
+		                sprintf(stackFrame, "(%s)\t0x%s — %s + %d",
+		                        moduleName, addr, functionName, offset);
+		            }
+
+		            if (functionName)
+		            {
+		                free(functionName);
+		            }
+		            Log::Write(LogLevel_Warning, "Stack: %s", stackFrame);
+		            std::string frameStr(stackFrame);
+		            stackFrames.push_back(frameStr);
+		        }
+		        free(strs);
+
+		        return stackFrames;
+		    }
+		};
+#endif
+#endif
+#endif
 	} // namespace Internal
 } // namespace OpenZWave
 
 /* keep this outside of the namespace */
 #if (defined _WINDOWS || defined WIN32 || defined _MSC_VER) && (!defined MINGW && !defined __MINGW32__ && !defined __MINGW64__)
 #include <ctime>
-struct tm *localtime_r(time_t *_clock, struct tm *_result);
+struct tm *localtime_r(const time_t *_clock, struct tm *_result);
 #endif
 
 #endif