+++ dzVents/documentation/history.md
@@ -2,6 +2,7 @@
 - Add method setMode for evohome device
 - Add method incrementCounter for incremental counter
 - Prepared for Firebase notifications. Firebase (fcm) is the replacement for Google Cloud Messaging gcm)
+- fix wildcard device 
 
 [2.4.22]
 - selector.switchSelector method accepts levelNames